Add temporary dashboard pointing to old data
We decided to run the new job "in-place" to avoid some complexity and
have data available as soon as possible.  To make it easy to compare new
data with old data, I'm adding this copy of the dashboard that points to
a backup of the data as of 2024-08-05.  This will stay up until we
validate that the new data is good, at which point we can delete it and
the backup data folder in hdfs.

Bug: T342267
Change-Id: I1a41ef34e3ba9ff4a970c46ccecca082c2ec4584